# packages/api-contract/generated/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sadhana API",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/sleep-records": {
      "get": {
        "operationId": "SleepRecords_getSleepRecords",
        "description": "Existing records for today and the previous four days in Moscow, newest first.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/sleep-records/yesterday/check": {
      "get": {
        "operationId": "SleepRecords_checkYesterday",
        "description": "True only when the day before yesterday has bedtime and yesterday has wake time in Moscow calendar dates.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckYesterdayData"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/sleep-records/{date}": {
      "get": {
        "operationId": "SleepRecords_getSleepRecord",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CalendarDate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyEntry"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "SleepRecords_putSleepRecord",
        "description": "Creates or replaces sleep data. Duration uses the previous day's bedtime and this day's wake time.",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CalendarDate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyEntry"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SleepDataInput"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/sleep-records/{date}/habits/{habitKey}": {
      "patch": {
        "operationId": "SleepRecords_setDailyHabitMark",
        "description": "Creates the daily entry if absent. The habit key need not exist in the habit list.",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CalendarDate"
            }
          },
          {
            "name": "habitKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyEntry"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyHabitMarkInput"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "SleepRecords_removeDailyHabitMark",
        "description": "Removing an absent mark succeeds when the daily entry exists.",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CalendarDate"
            }
          },
          {
            "name": "habitKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyEntry"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "message",
          "code"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "CalendarDate": {
        "type": "string",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
      },
      "CheckYesterdayData": {
        "type": "object",
        "required": [
          "hasData"
        ],
        "properties": {
          "hasData": {
            "type": "boolean"
          }
        }
      },
      "DailyEntry": {
        "type": "object",
        "required": [
          "id",
          "date",
          "sleep",
          "habits"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/CalendarDate"
          },
          "date": {
            "$ref": "#/components/schemas/CalendarDate"
          },
          "sleep": {
            "$ref": "#/components/schemas/SleepData"
          },
          "habits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HabitCheck"
            }
          }
        }
      },
      "DailyHabitMarkInput": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "boolean"
          }
        }
      },
      "HabitCheck": {
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "boolean"
          }
        }
      },
...
      "SleepData": {
        "type": "object",
        "required": [
          "bedtime",
          "wakeTime",
          "napDuration",
          "duration"
        ],
        "properties": {
          "bedtime": {
            "type": "string",
            "nullable": true,
            "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}$"
          },
          "wakeTime": {
            "type": "string",
            "nullable": true,
            "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}$"
          },
          "napDuration": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "duration": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0,
            "description": "Computed from the previous day's bedtime, this day's wake time, and nap duration. Null when either time is absent."
          }
        }
      },
      "SleepDataInput": {
        "type": "object",
        "required": [
          "napDuration"
        ],
        "properties": {
          "bedtime": {
            "type": "string",
            "nullable": true,
            "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}$"
          },
          "wakeTime": {
            "type": "string",
            "nullable": true,
            "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}$"
          },
          "napDuration": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0,
            "description": "Zero is stored as null."
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "Bearer"
      }
    }
  }
}